move legacy path mfa tests into a dedicated module
