+++ crates/defguard_proxy_manager/src/tests/proxy_manager/handler/mod.rs
@@ -4,6 +4,7 @@ mod acme;
 mod enrollment;
 mod lifecycle;
 mod mfa;
+mod mfa_legacy;
 mod oidc;
 mod password_reset;
 mod polling;
